AppRationalization/backend/requirements.txt
Flask
Flask-CORS>=3.0.10,<4
Flask-SQLAlchemy
python-dotenv
pypdf
pandas
openpyxl
numpy
SQLAlchemy
Werkzeug
requests
setuptools
wheel
packaging
PyMuPDF

# Optional: PostgreSQL support
# Uncomment the line below if using PostgreSQL as your database
# psycopg2-binary==2.9.6